+++ app/assets/stylesheets/application.scss
@@ -368,3 +368,9 @@ height: 300px;
   color: white;
   text-decoration: none;
 }
+
+.category_btn{
+  margin-right: 10px;
+  margin-bottom: 10px;
+  float: right;
+}
