+++ app/helpers/admins/categories_helper.rb
@@ -1,0 +1,2 @@
+module Admins::CategoriesHelper
+end
