+++ app/views/admins/categories/index.html.erb
@@ -1,0 +1,30 @@
+<div class= "container">
+    <h2>カテゴリ一覧</h2>
+    <%= link_to "カテゴリ作成", g_category_admins_categories_path, method: :post, class: "btn btn-primary category_btn" %>
+    <table class="table table">
+        <thead>
+            <tr>
+                <th>ID</th>
+                <th>ぐるなびID</th>
+                <th>カテゴリ名</th>
+                <th>更新日</th>
+            </tr>
+        </thead>
+        <tbody>
+            <% @categories.each do |c| %>
+                <tr>
+                    <td>
+                        <%= c.id %>
+                    </td>
+                    <td>
+                        <%= c.code %>
+                    </td>
+                    <td>
+                        <%= c.name %>
+                    </td>
+                    <td><h6 style="color: #C0C0C0;"><%= c.updated_at.strftime("%Y/%m/%d %H:%M") %></h6></td>
+                </tr>
+                <% end %>
+        </tbody>
+    </table>
+</div>
