+++ app/views/admins/chat_rooms/index.html.erb
@@ -1,34 +1,35 @@
-<h2>チャット一覧</h2>
-<table class="table table">
-       <thead>
-        <tr>
-            <th>ID</th>
-            <th>ユーザ</th>
-            <th>お店</th>
-            <th>更新日</th>
-        </tr>
-       </thead>
-       <tbody>
-        <% @chat_rooms.each do |cr| %>
+<div class= "container">
+    <h2>チャット一覧</h2>
+    <table class="table table">
+        <thead>
             <tr>
-                <td>
-                    <%= link_to cr.id, chat_room_path(cr.id)  %>
-                </td>
-                <td>
-                    <%= link_to admins_user_path(cr.user.id) do %>
-                        <%= attachment_image_tag cr.user, :user_image, format: 'jpeg', class: "round_image", fallback: "no_image.jpg", size: "30x30" %>
-                        <%= cr.user.full_name %>
-                    <% end %>
-                </td>
-                <td>
-                    <%= link_to admins_shop_path(cr.shop.id) do %>
-                        <%= attachment_image_tag cr.shop, :main_image, format: 'jpeg', class: "round_image", fallback: "no_image.jpg", size: "30x30" %>
-                        <%= cr.shop.name %>
-                    <% end %>
-                </td>
-                <td><h6 style="color: #C0C0C0;"><%= cr.updated_at.strftime("%Y/%m/%d %H:%M") %></h6></td>
+                <th>ID</th>
+                <th>ユーザ</th>
+                <th>お店</th>
+                <th>更新日</th>
             </tr>
-            <% end %>
-       </tbody>
-      
-</table>
+        </thead>
+        <tbody>
+            <% @chat_rooms.each do |cr| %>
+                <tr>
+                    <td>
+                        <%= link_to cr.id, chat_room_path(cr.id)  %>
+                    </td>
+                    <td>
+                        <%= link_to admins_user_path(cr.user.id) do %>
+                            <%= attachment_image_tag cr.user, :user_image, format: 'jpeg', class: "round_image", fallback: "no_image.jpg", size: "30x30" %>
+                            <%= cr.user.full_name %>
+                        <% end %>
+                    </td>
+                    <td>
+                        <%= link_to admins_shop_path(cr.shop.id) do %>
+                            <%= attachment_image_tag cr.shop, :main_image, format: 'jpeg', class: "round_image", fallback: "no_image.jpg", size: "30x30" %>
+                            <%= cr.shop.name %>
+                        <% end %>
+                    </td>
+                    <td><h6 style="color: #C0C0C0;"><%= cr.updated_at.strftime("%Y/%m/%d %H:%M") %></h6></td>
+                </tr>
+                <% end %>
+        </tbody>
+    </table>
+</div>
