+++ app/views/admins/homes/top.html.erb
@@ -1,8 +1,5 @@
-
-<%= link_to "カテゴリ作成", g_category_admins_shops_path, method: :post %>
-
-<div calss="container">
-<h3 class="heading">管理者画面</h3>
+<div class= "container">
+    <h3 class="heading">管理者画面</h3>
     <div class="row justify-content-center">
         <div class= "card admin-content admin-users">
             <div class="card-body">
@@ -49,7 +46,7 @@
                 <p>カテゴリ数</p>
                 <i class="fas fa-utensils fa-4x dashbord-icon"></i>
             </div>
-            <%= link_to admins_users_path, class:"card-footer text-center more-info" do %>
+            <%= link_to admins_categories_path, class:"card-footer text-center more-info" do %>
             <p>
             詳しい情報へ
             <span><i class="fas fa-arrow-alt-circle-right"></i></span>
